refactor:ListDto 스트림 형식 수정
